+++ es67x/src/main/scala/tech/beshu/ror/es/handler/CurrentUserMetadataRequestHandler.scala
@@ -44,12 +44,6 @@ class CurrentUserMetadataRequestHandler(engine: Engine,
   extends Logging {
 
   def handle(request: RequestContext.Aux[CurrentUserMetadataRequestBlockContext] with EsRequest[CurrentUserMetadataRequestBlockContext]): Task[Unit] = {
-    import tech.beshu.ror.accesscontrol.request.RequestContextOps._
-    engine.secondAccessControl
-    request.impersonateAs match {
-      case Some(impersonateAs) =>
-      case None =>
-    }
     engine.accessControl
       .handleMetadataRequest(request)
       .map { r => commitResult(r.result, request) }
