+++ es714x/src/main/scala/tech/beshu/ror/es/IndexLevelActionFilter.scala
@@ -32,6 +32,7 @@ import org.elasticsearch.transport.RemoteClusterService
 import tech.beshu.ror.accesscontrol.matchers.UniqueIdentifierGenerator
 import tech.beshu.ror.boot.RorSchedulers.Implicits.mainScheduler
 import tech.beshu.ror.boot._
+import tech.beshu.ror.boot.engines.Engines
 import tech.beshu.ror.es.handler.AclAwareRequestFilter
 import tech.beshu.ror.es.handler.AclAwareRequestFilter.EsContext
 import tech.beshu.ror.es.handler.response.ForbiddenResponse.{createRorNotReadyYetResponse, createRorStartingFailureResponse}
@@ -116,9 +117,9 @@ class IndexLevelActionFilter(clusterService: ClusterService,
         logger.warn(s"[${esContext.requestId}] Cannot handle the request ${esContext.channel.request().path()} because ReadonlyREST hasn't started yet")
         esContext.listener.onFailure(createRorNotReadyYetResponse())
       case RorInstanceStartingState.Started(instance) =>
-        instance.mainEngine match {
-          case Some(engine) =>
-            handleRequest(engine, esContext)
+        instance.engines match {
+          case Some(engines) =>
+            handleRequest(engines, esContext)
           case None =>
             logger.warn(s"[${esContext.requestId}] Cannot handle the request ${esContext.channel.request().path()} because ReadonlyREST hasn't started yet")
             esContext.listener.onFailure(createRorNotReadyYetResponse())
@@ -129,9 +130,9 @@ class IndexLevelActionFilter(clusterService: ClusterService,
     }
   }
 
-  private def handleRequest(engine: Engine, esContext: EsContext): Unit = {
+  private def handleRequest(engines: Engines, esContext: EsContext): Unit = {
     aclAwareRequestFilter
-      .handle(engine, esContext)
+      .handle(engines, esContext)
       .runAsync {
         case Right(_) =>
         case Left(ex) => esContext.listener.onFailure(new Exception(ex))
