+++ es714x/src/main/scala/tech/beshu/ror/es/handler/RegularRequestHandler.scala
@@ -52,7 +52,8 @@ class RegularRequestHandler(engine: Engine,
   extends Logging {
 
   def handle[B <: BlockContext : BlockContextUpdater](request: RequestContext.Aux[B] with EsRequest[B]): Task[Unit] = {
-    engine.accessControl
+    engine
+      .accessControl
       .handleRegularRequest(request)
       .map { r =>
         threadPool.getThreadContext.stashAndMergeResponseHeaders(esContext).bracket { _ =>
