+++ eshome/plugins/readonlyrest/plugin-descriptor.properties
@@ -55,7 +55,7 @@ java.version=1.8
 #
 # 'elasticsearch.version' version of elasticsearch compiled against
 
-elasticsearch.version=7.7.1
+elasticsearch.version=7.9.3
 
 # ES 2.x compatibility
 #site=false
