+++ proxy/src/main/scala/tech/beshu/ror/proxy/es/ProxyIndexLevelActionFilter.scala
@@ -14,6 +14,7 @@ import org.elasticsearch.common.settings.Settings
 import org.elasticsearch.tasks.Task
 import org.elasticsearch.threadpool.ThreadPool
 import tech.beshu.ror.accesscontrol.matchers.UniqueIdentifierGenerator
+import tech.beshu.ror.boot.engines.Engines
 import tech.beshu.ror.boot.{Engine, Ror, RorInstance, RorMode, StartingFailure}
 import tech.beshu.ror.es.handler.AclAwareRequestFilter
 import tech.beshu.ror.es.handler.AclAwareRequestFilter.EsContext
@@ -85,7 +86,7 @@ class ProxyIndexLevelActionFilter private(rorInstance: RorInstance,
                             channel: ProxyRestChannel): Unit = {
     aclAwareRequestFilter
       .handle(
-        engine,
+        Engines(engine, None), // todo: no second engine handling
         EsContext(
           channel,
           task,
